Add CMake preset support
With this we are now able to open this project in
VSCode and QtCreator and other IDEs that support
CMakePresets.

Add launch.json can can be further edited to also launch
godot editor

Add tasks.json for the build step. This can also be used
in combination with TaskRunner
https://marketplace.visualstudio.com/items?itemName=SanaAjani.taskrunnercode
for on click tasks (like formatting or custom deploy scripts). I use this
extensivly in ScreenPlay see
https://gitlab.com/kelteseth/ScreenPlay/-/blob/master/.vscode/tasks.json?ref_type=heads

Add godot.natvis for better debugger support

Add extensions.json that recommends cmake and cpp support for now. We
maybe should also add godot related extentions here.